add create like api, add delete like api, update like create service, update like delete service

+++ photobatle/service/like/Delete.py
@@ -9,8 +9,27 @@ class DeleteLikeService(Service):
     photo_id = forms.IntegerField()
     user_id = forms.IntegerField()
 
+    @property
+    def validate_photo_id(self):
+        try:
+            models.Photomodels.Photo.objects.get(pk=self.cleaned_data['photo_id'])
+        except Exception:
+            raise Exception(f"Incorrect photo_id value")
+        return True
+
+    @property
+    def get_like(self):
+        try:
+            models.Likemodels.Like.objects.get(photo_id=self.cleaned_data['photo_id'],
+                                               user_id=self.cleaned_data['user_id'])
+            return True
+        except Exception:
+            raise Exception(f"Photo dont have like from this user")
+
     def process(self):
-        like = models.Likemodels.Like.objects.get(photo_id=self.cleaned_data['photo_id'],
-                                                  user_id=self.cleaned_data['user_id'])
-        like.delete()
-        return (models.Photomodels.Photo.objects.get(pk=self.cleaned_data['photo_id'])).slug
+        if self.validate_photo_id:
+            if self.get_like:
+                like = models.Likemodels.Like.objects.get(photo_id=self.cleaned_data['photo_id'],
+                                                          user_id=self.cleaned_data['user_id'])
+                like.delete()
+                return (models.Photomodels.Photo.objects.get(pk=self.cleaned_data['photo_id'])).slug
